Moved the VGA driver to basic_drivers Added port_extensions for reading and writing whole buffers
